requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.7.1
Django==3.1.7
django-braces==1.14.0
django-debug-toolbar==3.2
django-oauth-toolkit==1.4.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
flower==0.9.7
hiredis==1.1.0
humanize==3.3.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
msgpack==1.0.2
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-django==4.1.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
zope.interface==5.2.0
